+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/politics/NationLeader.java
@@ -24,12 +24,6 @@ public class NationLeader extends Leader {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @OneToOne(mappedBy = "leader", cascade = {CascadeType.REFRESH,
-            CascadeType.DETACH,
-            CascadeType.MERGE,
-            CascadeType.PERSIST})
-    private Nation nation;
-
     @OneToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
             CascadeType.DETACH,
             CascadeType.MERGE,