+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/politics/military/SquadLeader.java
@@ -1,6 +1,7 @@
 package com.testament.veltahleon.model.entities.politics.military;
 
 import com.testament.veltahleon.abstraction.Leader;
+import com.testament.veltahleon.model.entities.politics.Rank;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
@@ -21,4 +22,11 @@ public class SquadLeader extends Leader {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @OneToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+            CascadeType.DETACH,
+            CascadeType.MERGE,
+            CascadeType.PERSIST})
+    @JoinColumn(name = "rank_id")
+    private Rank rank;
 }