+++ veltahleon/src/main/java/com/testament/veltahleon/repositories/politics/NationLeaderRepository.java
@@ -1,6 +1,5 @@
 package com.testament.veltahleon.repositories.politics;
 
-import com.testament.veltahleon.model.entities.history.library.Author;
 import com.testament.veltahleon.model.entities.politics.NationLeader;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -11,4 +10,5 @@ import java.util.Collection;
 public interface NationLeaderRepository extends JpaRepository<NationLeader, Long> {
 
     NationLeader findByName(String name);
+    Collection<NationLeader> findByBattle_Name(String name);
 }