+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/history/LetterService.java
@@ -9,7 +9,7 @@ public interface LetterService {
 
     Collection<Letter> getLettersWithPagination(int pageNumber, int numberOfRecords);
     Collection<Letter> getLetters();
-    //Collection<Letter> getLettersByLanguageName(String languageName);
+    Collection<Letter> getLettersByLanguageName(String languageName);
     Letter getLetterByID(Long id);
     Boolean deleteLetterByID(Long id);
     Letter saveLetter(Letter letter);