+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/politics/military/MilitaryLeaderService.java
@@ -8,7 +8,8 @@ public interface MilitaryLeaderService {
 
     Collection<MilitaryLeader> getMilitaryLeadersWithPagination(int pageNumber, int numberOfRecords);
     Collection<MilitaryLeader> getMilitaryLeaders();
-    Collection<MilitaryLeader> getMilitaryLeaderByNationName(String name);
+    Collection<MilitaryLeader> getMilitaryLeadersByNationName(String name);
+    Collection<MilitaryLeader> getMilitaryLeadersByBattleName(String name);
     MilitaryLeader getMilitaryLeaderByID(Long id);
     MilitaryLeader getMilitaryLeaderByName(String name);
     Boolean deleteMilitaryLeaderByID(Long id);